+++ packages/nimble-components/src/table-column/icon/tests/table-column-icon-matrix.stories.ts
@@ -73,9 +73,7 @@ const component = (): ViewTemplate => html`
     </${tableTag}>
 `;
 
-export const tableColumnIconThemeMatrix: StoryFn = createMatrixThemeStory(
-    component()
-);
+export const tableColumnIconThemeMatrix: StoryFn = createMatrixThemeStory(component());
 
 tableColumnIconThemeMatrix.play = async (): Promise<void> => {
     await Promise.all(